use Long ids

+++ hospital-jdbc/src/main/java/ir/maktab/hospital/dao/ReceptionDao.java
@@ -40,7 +40,7 @@ public class ReceptionDao implements BaseDao<ReceptionEntity> {
             ResultSet resultSet = ps.executeQuery();
 
             if (resultSet.next()) {
-                int id = resultSet.getInt(1);
+                Long id = resultSet.getLong(1);
                 String receptionDate = resultSet.getString(2);
                 String dischargeDate = resultSet.getString(3);
                 int patientId = resultSet.getInt(4);
